requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
sqlalchemy==2.0.35
rich==13.9.2